retrofitkit/drivers/raman/vendor_andor.py
import asyncio, time
from typing import Dict, Any
from retrofitkit.drivers.raman.base import RamanBase

# Placeholder: requires Andor SDK (Shamrock/Solis) python bindings
try:
    import andor  # fictitious placeholder
except Exception:
    andor = None

class AndorRaman(RamanBase):
    def __init__(self, cfg):
        self.t0 = time.time()
        self.dev = None

    async def read_frame(self) -> Dict[str, Any]:
        await asyncio.sleep(0.2)
        return {"t": time.time()-self.t0, "wavelengths":[532.0], "intensities":[995.0], "peak_nm":532.0, "peak_intensity":995.0}
